remove unused files, fix typing issues
