+++ vuu-ui/packages/vuu-table-extras/src/datagrid-configuration-ui/settings-panel/DatagridSettingsPanel.tsx
@@ -92,7 +92,7 @@ export const DatagridSettingsPanel = ({
 
   const tabstripProps: StackProps["TabstripProps"] = {
     activeTabIndex: selectedTabIndex,
-    enableRenameTab: false,
+    allowRenameTab: false,
     orientation: "vertical",
   };
