+++ vuu-ui/packages/vuu-ui-controls/src/drag-drop/index.ts
@@ -1,4 +1,4 @@
 export * from "./DragDropProvider";
 export * from "./dragDropTypes";
 export * from "./useDragDrop";
-export * from "./useDragDropNext";
+// export * from "./useDragDropNext";
