+++ vuu-ui/packages/vuu-utils/src/index.ts
@@ -15,6 +15,7 @@ export * from "./group-utils";
 export * from "./input-utils";
 export * from "./invariant";
 export * from "./json-utils";
+export * from "./keyboard-utils";
 export * from "./keyset";
 export * from "./logging-utils";
 export * from "./menu-utils";
